+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   - [composer](https://hub.docker.com/_/composer) with `--me`, `--pwd`, `--home`, `COMPOSER_HOME`
   - [node](https://hub.docker.com/_/node) with `--me`, `--pwd`, `--home`, `npm_config_cache`
 - Add `--volume` flag for a custom volume mount [#8](https://github.com/addshore/dockerit/issues/8)
+- `--port` flag can now be used multiple times
 
 ## 0.0.4 (7 February 2021)